Remove unused files. Sanitize code
